backend/requirements.txt
Django==4.2
djangorestframework==3.14.0
whitenoise==6.4.0
decouple==0.0.7
psycopg2-binary==2.9.6
python-dotenv==1.0.0
django-extensions==3.2.1
model-bakery==1.11.0
django-pytest==0.2.0
